requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.3
boto3==1.34.110
botocore==1.34.110
certifi==2024.2.2
charset-normalizer==3.3.2
configparser==4.0.2
crypto==1.4.1
hexdump==3.3
idna==3.7
jmespath==1.0.1
lxml==5.2.2
more-itertools==5.0.0
Naked==0.1.32
packaging==24.0
pymodbus==3.6.8
pyparsing==2.4.7
pyshark==0.6
python-dateutil==2.9.0.post0
python-evtx==0.7.4
PyYAML==6.0.1
requests==2.32.1
s3transfer==0.10.1
shellescape==3.8.1
six==1.16.0
soupsieve==2.5
termcolor==2.4.0
urllib3==2.2.1
zipp==1.0.0
